Moved "create_grid_with_halo" function into pad spatial module.
